requirements.txt
python-dotenv
beautifulsoup4
requests
streamlit
pandas
lxml
